v0.1.12: 05/07/2024
Started the process of converting hard coded positions and sizes of buttons, canvases, and labels into changeable variables. So far, I've done the Splits, Softs, Hards, and All buttons, the Random and In Order buttons, the Dealer Card Canvas, and the Player cards canvases.

Next steps are to place the Dealer Card, Dealer Suit, both Player Cards and Suits, in relative positions of the canvases.

Then work on making variables for the action buttons and all font sizes.

Also need to change the size of the button fonts.